Insert Method
Insert method takes in what ever value is passed into it's parameters and returns a  linked node.

first it checks if there is a value in the head poiinter, if there is we will set our value to next pointer.

If there is no value in the head pointer, we will add our value into the head pointer